requirements.txt
python-dotenv
google-genai
neo4j
pydantic